requirements.txt
numpy==2.2.6
onnxruntime==1.15.0
opencv-python==4.12.0.88       # 若你需要 GUI 顯示功能，選這個
# opencv-python-headless==4.10.0.84  # ← 若你只在 server/headless 執行時用這個，不能同時裝兩個

Pillow==11.3.0
pycocotools==2.0.10
setuptools==65.5.0
torch==2.7.1
torchvision==0.22.1
transformers==4.53.2
scikit-learn==1.5.0             # ✅ 加上明確版本號（你可以調整版本需求）

git+https://github.com/openai/CLIP.git

